+++ .gitignore
@@ -153,7 +153,6 @@ cython_debug/
 
 # Contains PII
 Forecasts/*
-utils/private.json
-utils/private.py
+utils/private/*
 gui/icons/*.png
-utils/json/raw_forecast_test.json
+utils/json/test_data/raw_forecast.json
