+++ gui/notebooks/forecast_viewer.py
@@ -4,13 +4,13 @@ from tkinter import messagebox
 from gui.empyrean.datetime import EmpyreanDateTime
 from gui.empyrean.labelframe import LabelFrame
 from gui.empyrean.notebook import Notebook
+from gui.widget_enum import WidgetType
 from gui.windows.request_manager import *
 from utils.download.download_status import *
-from utils.forecast.forecast_type import *
-from utils.forecast.location import *
-from utils.gridplacement import GridPlacement
-from utils.json.forecast import Forecast
-from utils.WidgetEnum import WidgetType
+from utils.structures.forecast.forecast import Forecast
+from utils.structures.forecast.forecast_type import *
+from utils.structures.grid_placement import GridPlacement
+from utils.structures.location.location import Location
 
 
 class ForecastViewer_Notebook(Notebook):
