--- utils/dataclasses/forecast_test_data.json
@@ -1,62 +1,0 @@
-[
-    {
-        "forecast_type": "hourly",
-        "units": "us",
-        "info" : {
-            "generatedAt": {
-                "date": "2024-12-23",
-                "time": "03:11"
-            },
-            "updateTime": {
-                "date": "2024-12-23",
-                "time": "02:53"
-            },
-            "validTimes": {
-                "date": "2024-12-23",
-                "time": "20:00"
-            },
-            "elevation": {
-                "unitCode": "wmoUnit:m",
-                "value": 1553.8704
-            },
-            "short" : "Cloudy",
-            "long" : "Rain showers before 7pm, then showers and thunderstorms between 7pm and 1am, then areas of fog and a chance of rain showers. Mostly cloudy. Low around 42, with temperatures rising to around 44 overnight. South wind 3 to 8 mph. Chance of precipitation is 100%. New rainfall amounts between a tenth and quarter of an inch possible."
-        },
-        "subdata": [
-            {
-                "name" : "2024-12-22",
-                "open" : true,
-                "subdata" : [
-                    {
-                        "name" : "20:00",
-                        "open" : true,
-                        "measurements" : [
-                            {
-                                "name" : "Temperature",
-                                "value" : "42 F"
-                            },
-                            {
-                                "name" : "Rain Chance",
-                                "value" : "5 %"
-                            }
-                        ]
-                    },
-                    {
-                        "name" : "21:00",
-                        "open" : true,
-                        "measurements" : [
-                            {
-                                "name" : "Temperature",
-                                "value" : "41 F"
-                            },
-                            {
-                                "name" : "Rain Chance",
-                                "value" : "12 %"
-                            }
-                        ]
-                    }
-                ]
-            }
-        ]
-    }
-]
