+++ utils/download/html_status.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 class HTMLStatusCode(Enum):
     OK                  = 200
     NOT_FOUND           = 404
