+++ utils/structures/forecast/elevation.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Elevation():
+    unitCode    : str
+    value       : float
