+++ utils/structures/forecast/forecast_type.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 class ForecastType(Enum):
     HOURLY      = 'hourly'
     EXTENDED    = 'extended'
