+++ utils/structures/location/location.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from typing import Any
 
+
 @dataclass
 class Position():
     lattitude   : str
